requirements.txt
# Core dependencies
PyYAML>=6.0
psycopg2-binary>=2.9.0
requests>=2.28.0
chardet>=5.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Optional performance improvements
ujson>=5.0.0